+++ app/members/urls.py
@@ -9,6 +9,5 @@ router.register(r'user', views.UserModelViewSet, basename='UserModel')
 urlpatterns_members = [
     path('kakaoToken/', apis.KakaoJwtTokenView.as_view()),
     path('facebookToken/', apis.FacebookJwtToken.as_view()),
-    path('', apis.SignUpView.as_view()),
 ]
 urlpatterns_members += router.urls
